Rebased merge master
HTML Output

This creates the HTML File of the filtered timeline.

IEvents to Nodes and Connecting

This is connecting the checkbox to the function. I fixed the download button so it is uniform with the others. I also changed the loop through IEvents to Nodes

Adds multiselect to event type filter and fixes combobox duplications

- Adds the ability to multiselect eventTypes via the UI
- Fixes bug in event type where options would be duplicated upon opening
- fixes duplicate "hours" in timeline scaling combobox

- removes expired TODO and FIXME

- fixed error where datetime and event type would destroy nodes instead of hiding
- modified unit tests to check for consistent number of nodes during filtering
- Added "blank" filter handling for event name and event type
- added functionality to restore nodes previously hidden when removing filters

Connect the CsvIO to UI checkbox

SaveFileDialog

User can now name the html/csv and choose the location of the files

Add script API endpoint adds a script to the database

Changed the add scripts endpoint to accept encoded scripts and added endpoint to get encoded scripts

Logs are printed in both the console and in a log file found in the same location as the exe